requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3